move gitlab config to separate folder
