+++ src/app/app.module.ts
@@ -23,8 +23,8 @@ import { persistState } from '@datorama/akita';
 import { AkitaNgDevtools } from '@datorama/akita-ngdevtools';
 import { environment } from '../environments/environment';
 import { AppComponent } from './app.component';
-import { GitlabConfigDialogComponent } from './gitlab-config-dialog/gitlab-config-dialog.component';
-import { GitlabConfigItemComponent } from './gitlab-config-item/gitlab-config-item.component';
+import { GitlabConfigDialogComponent } from './gitlab-config/gitlab-config-dialog/gitlab-config-dialog.component';
+import { GitlabConfigItemComponent } from './gitlab-config/gitlab-config-item/gitlab-config-item.component';
 import { HttpErrorInterceptor } from './http-error.interceptor';
 import { AKITA_PERSIST_STORAGE } from './persist-state.token';
 import { QueryFormComponent } from './query-form/query-form.component';
