+++ src/app/gitlab-config/gitlab-config-dialog/gitlab-config-dialog.component.spec.ts
