+++ src/app/gitlab-config/gitlab-config-item/gitlab-config-item.component.scss
