+++ src/app/query-form/query-form.component.html
@@ -8,9 +8,17 @@
     <mat-error *ngIf="!projectSelected">Please select at least one GitLab project</mat-error>
   </div>
   <div>
-    <button mat-raised-button color="primary" [disabled]="!projectSelected || queryForm.pristine || queryForm.invalid || resultLoading">
+    <button
+      mat-raised-button
+      color="primary"
+      *ngIf="!resultLoading; else stopSearching"
+      [disabled]="!projectSelected || queryForm.pristine || queryForm.invalid"
+    >
       <mat-icon>search</mat-icon>&nbsp;Search
     </button>
+    <ng-template #stopSearching>
+      <button mat-raised-button color="warn" type="button" (click)="stop.emit()"><mat-icon>stop</mat-icon>&nbsp;Stop</button>
+    </ng-template>
   </div>
 </form>
 <mat-progress-bar mode="indeterminate" *ngIf="resultLoading"></mat-progress-bar>
