+++ src/app/query-form/query-form.component.ts
@@ -29,6 +29,7 @@ export class QueryFormComponent {
   }
 
   @Output() query = new EventEmitter<string>();
+  @Output() stop = new EventEmitter<void>();
 
   submit(): void {
     this.query.emit(this.queryForm.value.query);
